requirements.txt
coverage
pip>20
numpy>=1.15.4
pandas
python-libsbml
simplesbml
skia-python
ipython
Pillow
setuptools>=42
openpyxl
opencv-python
networkx
matplotlib
